Passing VectorNormals tests
A few issues addressed:

- There were what looked a lot like finite-precision
errors happening in the normals funcs when using
cartesian_positions as inputs. Fixed by now using
float64 when performing that arithmetic

- Signs were getting flipped in the octencoded
vertex normal values. This went away when I implemented
the TODO item in oct_encode

- Tests now comparing vertex normals with an allowable
error of tolerance to account for accuracy lost in
the octencoding/decoding